### 2021-03-15 #3112 @ulf1
Hello @arfon I moved the repository into a new organisation but couldn't change URL anymore. The new location is https://github.com/satzbeleg/bwsample
Is there a way to change the path?
