+++ bwsample/__init__.py
@@ -1,6 +1,7 @@
-__version__ = '0.6.1'
+__version__ = '0.6.2'
 
 from .sampling import sample
 from .counting import count
 from .ranking import rank
 from .utils import (to_scipy, add_dok, adjustscore)
+from .maxdiff import scoring_orme
