### 2021-04-01 95b217b
bugfix, set to 0 if counted less
