### 2021-04-01 f985a08
multiply with a scalar, gt instead of lt
