+++ bwsample/ranking.py
@@ -316,10 +316,10 @@ def maximize_hoaglinapprox(cnt: scipy.sparse.csr_matrix,
     # compute Hoaglin's Approximation for DoF=0
     P = np.sqrt(X2)
     P.data = np.power(0.1, (P.data + 1.37266) / 2.13161)
-    # P.data = P.data * 4.405087805849058
+    P.data = P.data * 4.405087805849058
     P.data = np.maximum(0.0, np.minimum(1.0, P.data))
     # only if Nij>Nji
-    P = P.multiply(cnt < cnt.T)
+    P = P.multiply(cnt > cnt.T)
 
     # sum rows in DoK matrix
     metrics = np.array(P.sum(axis=1).flatten())[0]
