+++ .github/FUNDING.yml
@@ -1,2 +1,4 @@
 github: [ulf1]
-issuehunt: satzbeleg/bwsample
+patreon: ulfhamster
+ko_fi: ulfhamster
+
