+++ requirements.txt
@@ -1,4 +1,4 @@
 # public packages (see setup.py)
-numpy>=1.19.5
-scipy>=1.5.4
+numpy>=1.19.5,<2
+scipy>=1.5.4,<2
 scikit-learn>=0.24.1
