### 2021-04-23 e9671dd
remove pandoc from setup.py
