+++ .github/workflows/syntax-and-unit-tests.yml
@@ -7,16 +7,23 @@ jobs:
 
     runs-on: ubuntu-18.04
 
+    strategy:
+      matrix:
+        python-version: ['3.6', '3.7', '3.8', '3.x', 'pypy3']
+    
+    name: Python ${{ matrix.python-version }} Tests
+
     steps:
     - uses: actions/checkout@v1
-    - name: Set up Python 3.6
+    - name: Setup python
       uses: actions/setup-python@v1
       with:
-        python-version: 3.6
+        python-version: ${{ matrix.python-version }}
+        architecture: x64
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements.txt
+        python setup.py install
         pip install -r requirements-dev.txt
     - name: Lint with flake8
       run: |
