### 2021-06-01 #3324 @whedon
[REVIEW]: `bwsample`: Processing Best-Worst Scaling data
**Submitting author:** @ulf1 (<a href="http://orcid.org/0000-0002-0440-4868">Ulf Hamster</a>)
**Repository:** <a href="https://github.com/satzbeleg/bwsample" target ="_blank">https://github.com/satzbeleg/bwsample</a>
**Version:** v0.6.9
