### 2021-06-07 792dc76
minmax scaling added to adjustscores function
