+++ bwsample/sampling.py
@@ -1,5 +1,4 @@
 import numpy as np
-import random
 import warnings
 from typing import List, Optional
 
