### 2021-06-08 2379f88
deepcode badge removed because service fades out soon, or resp becomes a paid service by the mothership
