### 2021-06-20 #3324 @ejhigson
> @ejhigson & @jakryd, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
>
> 1. Make sure you're logged in to your GitHub account
> 2. Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations
>
> The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let @mikldk know.

Hi @mikldk - I'm trying to update my checklist but I can't edit it. When I click on https://github.com/openjournals/joss-reviews/invitations I get a message saying "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." Do you know what has gone wrong here? Please could you perhaps resend my invitation to do the review? Thank you!
