### 2021-06-28 #3324 @jakryd
Sorry for the delay, @mikldk. I'm almost over with my review.

@ulf1 -- Have you compared how your methods perform in comparison to other commonly-used packages?
