### 2021-06-28 #3324 @ulf1
> Sorry for the delay, @mikldk. I'm almost over with my review.
>
> @ulf1 -- Have you compared how your methods perform in comparison to other commonly-used packages?

To my surprise, I haven't found a package that suited our needs. There exists a commercial software vendor (Sawtooth Software) who offers Best-Worst-Scaling surveys since the 80s/90s. It looks that they are related to the research community about BWS in the 90s.
Our UI (see page 2 in https://osf.io/qkxej/) and Sawtooth's UI (see https://sawtoothsoftware.com/maxdiff) will produce the same kind of data but that's basically all. The 80s/90s BWS method is classic statistics, e.g. items will be shown several times  over and over again. The assumptions is that the number of items is somewhat limited.
In our use case we have billions of items to assess (sentence examples) but still the same human resource limitations. As a result we had the idea to transform the BWS data in paired comparison data, and logical infer even more pairs (see Ch.4 in https://osf.io/qkxej/). This was implemented in `bwsample.count`. This approach requires also a certain kind of sampling strategy (see Ch.5 in https://osf.io/qkxej/) what is available in `bwsample.sample`.
Regarding the ranking of the (big) sparse matrix with paired comparisons (database with up to billions^2 pairs), i prefer the most simple approaches (e.g. `bwsample.rank(..., "ratio")`) because the **update** can be done with linear complexity in production. The package also includes more theoretical sound approaches (e.g. BTL Maximum Likelihood). The little experiment in Ch.7 (https://osf.io/ev7fw/) shows that the BTL models basically produce the same rankings. However, the problem is that approaches like BTL-model, Eigenvector, etc. require a complete re-computation that is `O(n^2)` complex what is infeasible in our production scenario.
