### 2021-07-01 a8b22a5
Create codeql-analysis.yml
