+++ bwsample/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.6.6'
+__version__ = '0.6.7'
 
 from .sampling import sample
 from .counting import count
