### 2021-07-18 #3324 @ejhigson
Hi @ulf1 - congratulations on a nice software package! Everything ran with no issues for me. I only have a couple of very minor comments on the paper:

1) One of the criteria is to state "how this software compares to other commonly-used packages" - please can you comment on this? If you don't know of any other best-worst sampling packages then just stating that is fine I think.

2) I saw a few small typos:
L16 “software feature” -> “software features”
L19 “items has to be shown” -> “items have to be shown”
L20 “reasonable amounts counting” -> “reasonable amounts of counting
