### 2021-07-18 #3324 @ulf1
@ejhigson
1. No, I don't know comparable software packages. See also the comment above https://github.com/openjournals/joss-reviews/issues/3324#issuecomment-869620104
2. Thank you for spotting the typos.
