### 2021-07-18 satzbeleg/bwsample#72 @ulf1
Correct the typos
https://github.com/openjournals/joss-reviews/issues/3324#issuecomment-882063865
