### 2021-07-19 #3324 @ejhigson
> @ejhigson
>
> 1. No, I don't know comparable software packages. See also the comment above [[REVIEW]: `bwsample`: Processing Best-Worst Scaling data #3324 (comment)](https://github.com/openjournals/joss-reviews/issues/3324#issuecomment-869620104)
> 2. Thank you for spotting the typos.

Thank you very much for this, and sorry I had missed the above comment! I have now completed my review.
