### 2021-07-26 #3324 @mikldk
@ejhigson, @jakryd:  Can you confirm that you have finished the review and recommend that this paper is now published once satzbeleg/bwsample#72 is closed?

@ulf1 Please ping me once satzbeleg/bwsample#72 is closed and that you had a final read though of the paper, checking language etc. and a final check of the proofs with `@whedon generate pdf`.
