### 2021-07-26 23b36c7
fix #72 typos corrected
