### 2021-08-12 #3324 @mikldk
@ulf1 I see that you follow semantic versioning, but have both a `0.6.7` and a `0.6.7-joss`. I can see that only the `README.md` was changed (together with the paper). Please bump the version to e.g. `0.6.7.1` or `0.6.8` -- or else it would be inconsistent if we here point to `0.6.7`, but in reality if should be `0.6.7-joss`. Sorry for being pedantic.

Also, please make the names at Zenodo and the paper the same (middle name initial).
