### 2021-08-13 #3324 @ulf1
@mikldk I bumped the version to 0.6.8
- https://github.com/satzbeleg/bwsample/releases/tag/0.6.8
- https://pypi.org/project/bwsample/0.6.8/
- https://doi.org/10.5281/zenodo.5196171

Be careful, it's Friday the 13th 👻
