+++ bwsample/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.6.8'
+__version__ = '0.6.9'
 
 from .sampling import sample
 from .counting import count
