### 2021-08-22 #3324 @ulf1
@mikldk the title and author name are aligned now.
- https://doi.org/10.5281/zenodo.5233945
- https://github.com/satzbeleg/bwsample/blob/main/.zenodo.json
- https://github.com/satzbeleg/bwsample/blob/main/docs/paper.md
