### 2021-08-23 #3324 @danielskatz
👋 @ulf1 - As the AEiC on duty this week, I've suggested some changes to the paper in https://github.com/satzbeleg/bwsample/pull/75 - please merge this, or let me know what you disagree with.  Once we do this, we can proceed to final publication.
